requirements.txt
aniso8601==9.0.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
firebase-admin==6.1.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-cloud-tasks==2.13.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.5
packaging==23.1
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
stripe==5.4.0
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
